backend/requirements.txt
Django
python-dotenv
psycopg2-binary
django-cors-headers
djangorestframework
